requirements.txt
cairosvg==2.7.1
Pillow==11.0.0